+++ tomcat/src/main/java/nextstep/jwp/controller/LoginController.java
@@ -3,10 +3,12 @@ package nextstep.jwp.controller;
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.dto.LoginRequest;
 import nextstep.jwp.exception.UnauthorizedException;
+import nextstep.jwp.http.Headers;
+import nextstep.jwp.http.QueryStringConverter;
+import nextstep.jwp.http.RequestEntity;
+import nextstep.jwp.http.ResponseEntity;
 import nextstep.jwp.model.User;
-import nextstep.jwp.support.QueryStringConverter;
-import nextstep.jwp.support.Resource;
-import nextstep.jwp.support.View;
+import nextstep.jwp.support.*;
 import org.apache.http.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
