+++ tomcat/src/main/java/nextstep/jwp/http/QueryStringConverter.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.support;
+package nextstep.jwp.http;
 
 import java.util.HashMap;
 import java.util.Map;
