+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -4,8 +4,11 @@ import nextstep.jwp.controller.Controller;
 import nextstep.jwp.exception.CustomNotFoundException;
 import nextstep.jwp.exception.FileAccessException;
 import nextstep.jwp.exception.UnauthorizedException;
-import nextstep.jwp.support.Resource;
-import nextstep.jwp.support.View;
+import nextstep.jwp.http.Headers;
+import nextstep.jwp.http.RequestEntity;
+import nextstep.jwp.http.RequestInfo;
+import nextstep.jwp.http.ResponseEntity;
+import nextstep.jwp.support.*;
 import org.apache.coyote.Processor;
 import org.apache.http.*;
 import org.slf4j.Logger;
