+++ tomcat/src/test/java/nextstep/jwp/controller/ResourceControllerTest.java
@@ -1,6 +1,9 @@
 package nextstep.jwp.controller;
 
 import nextstep.jwp.exception.CustomNotFoundException;
+import nextstep.jwp.http.Headers;
+import nextstep.jwp.http.RequestEntity;
+import nextstep.jwp.http.ResponseEntity;
 import org.apache.http.*;
 import org.junit.jupiter.api.Test;
