+++ tomcat/src/test/java/org/apache/http/HeadersTest.java
@@ -1,5 +1,6 @@
 package org.apache.http;
 
+import nextstep.jwp.http.Headers;
 import nextstep.jwp.support.View;
 import org.junit.jupiter.api.Test;
